Use a factory function instead of a constructor.
